% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{A function that creates a linear regression for a dataframe}
\usage{
linear_regression(df)
}
\description{
A function that creates a linear regression for a dataframe
}
\examples{
linear_regression()
}
\keyword{plot}
