% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_regression.R
\name{quadratic_regression}
\alias{quadratic_regression}
\title{A function that creates a quadratic regression for a dataframe}
\usage{
quadratic_regression(df)
}
\description{
A function that creates a quadratic regression for a dataframe
}
\examples{
quadratic_regression()
}
\keyword{plot}
