% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_regression.R
\name{which_regression}
\alias{which_regression}
\title{A function that helps user decide which regression is best for a dataframe}
\usage{
which_regression(df)
}
\description{
A function that helps user decide which regression is best for a dataframe
}
\examples{
which_regression()
}
\keyword{plot}
